server/requirements.txt
numpy<2.0
openai-whisper
websockets